## Linked issue (required)
<!-- Use a closing keyword so it shows up in Development -->
Fixes #81 

## What changed
- converted cvalue to double
- added it to fit as a minumul corelation value

## Why
- smoother fitting

## How to test
-  you'll see it

+++ inc/printing.h
@@ -39,7 +39,7 @@ void printspaces(int maxpos,
                  size_t mod,
                  size_t acolumns,
                  double avgPos,
-                 size_t correlationThreshold);
+                 double correlationThreshold);
 
 /** @brief Prints a formatted message to the standard error stream, using a
  printf-style format string and variable arguments.
